Move `NightscoutDTO` to mongodb package
